requirements.txt
python-docx>=0.8.11
pywin32>=306; sys_platform == 'win32'
pillow>=9.0.0
pyinstaller>=5.6.2 